Fix "Object should be of type List" errors when adding valid context menu items
This is stupid and should not be necessary, or even work but it does.
Apparently XBMC has some weird problem with single element collections.
It causes perfectly valid lists to (sometimes) fail when passing them to
add_item. Don't ask me why using context_menu_items in a log statement
changes anything, but it does. I can only imagine that internally XBMC
recognizes that it's a single item list and corrects itself.